add react-hooks package, pull reflective hooks out of controls
