+++ javascript/react-controls/src/controls/index.ts
@@ -3,7 +3,6 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 export * from './ColumnarMenu/index.js'
-export * from './ColumnInstruction.js'
 export * from './ColumnOrValueComboBox.js'
 export * from './ColumnSpread.js'
 export * from './ColumnValueDropdown.js'
@@ -15,4 +14,4 @@ export * from './MultiDropdown.js'
 export * from './ReadOnlyTextField.js'
 export * from './SortInstruction.js'
 export * from './TableColumnDropdown.js'
-export * from './TableDropdown/index.js'
+export * from './TableDropdown.js'
