+++ javascript/react-hocs/src/withInputColumnDropdown.tsx
@@ -21,7 +21,7 @@ import { memo, useCallback, useMemo } from 'react'
 import styled from 'styled-components'
 
 import type { HOCFunction } from './types.js'
-import { useTableColumnOptions } from '@data-wrangling-components/react-controls'
+import { useTableColumnOptions } from '@data-wrangling-components/react-hooks'
 /**
  * Higher order component generator to wrap a Step in the input column dropdown.
  * @param label - optional label to use for the dropdown instead of the default.
