+++ javascript/react-hooks/.npmignore
@@ -1,0 +1,2 @@
+.swc/
+tsconfig.json
