+++ javascript/react-hooks/docs/react-hooks.api.md
@@ -1,0 +1,29 @@
+## API Report File for "@data-wrangling-components/react-hooks"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+import type ColumnTable from 'arquero/dist/types/table/column-table';
+import type { IDropdownOption } from '@fluentui/react';
+import type { TableStore } from '@data-wrangling-components/core';
+import type { Value } from '@essex/arquero';
+
+// Warning: (ae-missing-release-tag) "useColumnValueOptions" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export function useColumnValueOptions(column: string | undefined, table: ColumnTable | undefined, values?: Value[], filter?: (value: Value) => boolean): IDropdownOption[];
+
+// Warning: (ae-missing-release-tag) "useTableColumnOptions" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public
+export function useTableColumnOptions(table: ColumnTable | undefined, filter?: (name: string) => boolean): IDropdownOption[];
+
+// Warning: (ae-missing-release-tag) "useTableOptions" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public
+export function useTableOptions(store?: TableStore): IDropdownOption[];
+
+// (No @packageDocumentation comment for this package)
+
+```
