+++ javascript/react-hooks/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "@data-wrangling-components/react-hocs",
+	"name": "@data-wrangling-components/react-hooks",
 	"version": "0.1.0",
 	"type": "module",
 	"main": "src/index.ts",
@@ -33,8 +33,6 @@
 	"dependencies": {
 		"@data-wrangling-components/core": "workspace:^",
 		"@data-wrangling-components/react-controls": "workspace:^",
-		"@data-wrangling-components/react-hooks": "workspace:^",
-		"@data-wrangling-components/react-verbs": "workspace:^",
 		"@essex/arquero": "^2.0.2",
 		"@essex/dataflow": "^0.1.1"
 	},
