+++ javascript/react-hooks/src/hooks.ts
@@ -1,0 +1,83 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+import type { Value } from '@essex/arquero'
+import type { IDropdownOption } from '@fluentui/react'
+import { op } from 'arquero'
+import type ColumnTable from 'arquero/dist/types/table/column-table'
+import { useMemo } from 'react'
+
+import type { TableStore } from '@data-wrangling-components/core'
+import { useSimpleDropdownOptions } from '@data-wrangling-components/react-controls'
+import { useEffect, useState } from 'react'
+
+/**
+ * Creates a list of dropdown options for the columns in a table
+ * @param table - the input table
+ * @returns a list of dropdown options
+ */
+export function useTableColumnOptions(
+	table: ColumnTable | undefined,
+	filter?: (name: string) => boolean,
+): IDropdownOption[] {
+	return useSimpleDropdownOptions(table?.columnNames(filter) || [])
+}
+
+export function useColumnValueOptions(
+	column: string | undefined,
+	table: ColumnTable | undefined,
+	values?: Value[],
+	filter?: (value: Value) => boolean,
+): IDropdownOption[] {
+	const vals = useMemo(() => {
+		if (!column || !table || !column || column.trim().length === 0) {
+			return []
+		}
+		const getFallback = () => {
+			const columnNamesArray: string[] = table
+				.columnNames()
+				.filter(e => e === column)
+
+			if (columnNamesArray.length !== 0) {
+				const result: any[] = table
+					.rollup({
+						[column]: op.array_agg_distinct(column),
+					})
+					.get(column, 0)
+
+				return result ?? []
+			}
+
+			return []
+		}
+		const list = values ? values : getFallback()
+		return filter ? list.filter(filter) : list
+	}, [column, table, values, filter])
+	return useSimpleDropdownOptions(vals)
+}
+
+/**
+ * Creates a list of dropdown options from the tables in a store
+ * TODO: for any given step, we should only show the tables created *prior* to this step,
+ * potentially via an optional filter callback on store.list.
+ * As it is, whenever the store is updated all the table dropdowns get the results.
+ * @param store -
+ * @returns
+ */
+export function useTableOptions(store?: TableStore): IDropdownOption[] {
+	// we won't actually get an updated store reference, so we'll track
+	// whether updates are needed using a change listener and flag
+	const [dirty, setDirty] = useState<boolean>(true)
+	const [list, setList] = useState<string[]>([])
+	useEffect(() => {
+		return store?.onChange(() => setTimeout(() => setDirty(true), 0))
+	}, [store, setDirty])
+	useEffect(() => {
+		if (dirty) {
+			setDirty(false)
+			setList(store?.list().sort() || [])
+		}
+	}, [store, dirty, setDirty, setList])
+	return useSimpleDropdownOptions(list)
+}
