+++ javascript/react-hooks/src/index.ts
@@ -2,4 +2,4 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-export * from './TableDropdown.js'
+export * from './hooks.js'
