+++ javascript/react-verbs/src/verbs/Pivot/Pivot.tsx
@@ -14,7 +14,7 @@ import styled from 'styled-components'
 import { LeftAlignedRow, useHandleDropdownChange } from '../../common/index.js'
 import { withLoadedTable } from '../../common/withLoadedTable.js'
 import type { StepComponentProps } from '../../types.js'
-import { useTableColumnOptions } from '@data-wrangling-components/react-controls'
+import { useTableColumnOptions } from '@data-wrangling-components/react-hooks'
 
 /**
  * Just the group/column/op inputs for an aggregation.
